BookShelf/src/bookshelf/model/User.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package bookshelf.model;

/**
 *
 * 
 * User osztály a felhasználó adatainak megadásához, user adatokkal és hozzájuk tartozó propertykkel
 */
public class User {
    private int userId;
    private String password;
    private String username;

    public void setUserId(int userId) {
        this.userId = userId;
    }

    public int getUserId() {
        return userId;
    }    

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
    
}
